NB1: Rework fstab files
* The stock fstab file from Nokia has way too many partitions, most
of which are useless to us. Some of them cause harmless warnings (elabel, systeminfo),
while others cause slot switch failures (xbl) when included in the fstab. To address this
fetch the fstab from the NB1 LineageOS tree and adjust it for recovery
usage.

* Also mount persist properly at /mnt/vendor/persist and create a symlink to at /persist
because TWRP doesn't support mounting partitions at /mnt